fixed issues with testrunner
