+++ testutils/helper/http_echo.go
@@ -1,0 +1,13 @@
+package helper
+
+const (
+	defaultHttpEchoImage = "kennship/http-echo:latest"
+	HttpEchoName         = "http-echo"
+	HttpEchoPort         = 3000
+
+)
+
+
+func NewEchoHttp(namespace string) (*TestContainer, error) {
+	return newTestContainer(namespace, defaultHttpEchoImage, HttpEchoName, HttpEchoPort)
+}
